changed fitness function to only return ints not longs. changed Weighted Roulette Selector to retrieve entrySets instead of keysets.
git-svn-id: file:///home/mast/dev/jgap/svn/trunk@25 3924feb9-9058-4ba4-b128-58a6af0bd8c6

+++ org/jgap/FitnessFunction.java
@@ -53,6 +53,6 @@ public interface FitnessFunction extends java.io.Serializable {
    *
    * @return The fitness rating of the given Chromosome.
    */
-  public long evaluate(Chromosome subject);
+  public int evaluate(Chromosome subject);
 }
